+++ client-next/codegen.ts
@@ -1,0 +1,15 @@
+import type { CodegenConfig } from '@graphql-codegen/cli';
+
+const config: CodegenConfig = {
+  overwrite: true,
+  schema: '../src/ext/graphql/transmodelapi/schema.graphql',
+  documents: 'src/**/*.{ts,tsx}',
+  generates: {
+    'src/gql/': {
+      preset: 'client',
+      plugins: [],
+    },
+  },
+};
+
+export default config;