+++ src/test/java/org/opentripplanner/mmri/TimeTest.java
@@ -78,8 +78,8 @@ public class TimeTest extends GtfsTest {
 
     Leg leg = itinerary.getLegs().toArray(new Leg[1])[0];
 
-    validateLeg(leg, 1388703780000L, 1388703840000L, "1g2", "1g1", null);
+    validateLeg(leg, 1388703660000L, 1388703720000L, "1g2", "1g1", null);
 
-    assertEquals("Stop 1g1 ~ BUS bus 0:03 0:04 ~ Stop 1g2 [C₁90]", itinerary.toStr());
+    assertEquals("Stop 1g1 ~ BUS bus 0:01 0:02 ~ Stop 1g2 [C₁90]", itinerary.toStr());
   }
 }